New tighter PAGENT rules for VSAM prep
To get ready for VSAM/CDC, the existing PAGENT rule for Db2 had to be
fixed.  It uses a DNS lookup to query the IP addresses of the ROKS
load balancer and adds them to an IP address group, which is then
referenced in the PAGENT rule.  The wide port range has been shut down
to just 443 as well.